refactor(phase6): extract clinician document workflows
