+++ backend/src/app/services/clinician_service.py
@@ -32,7 +32,7 @@ from app.core.exceptions import (
 from app.db.repositories import CareTeamRepository, ClinicianRepository, ClinicRepository
 from app.db.supabase_execute import execute_async
 from app.models.dashboard import DashboardSortBy, DashboardSortOrder, RiskLevel
-from app.models.enums import DocumentReviewStatus, UploaderRole
+from app.services.clinician_document_workflow_service import ClinicianDocumentWorkflowService
 
 logger = logging.getLogger(__name__)
 
@@ -45,6 +45,11 @@ class ClinicianService:
         self.clinic_repo = ClinicRepository(self)
         self.clinician_repo = ClinicianRepository(self)
         self.care_team_repo = CareTeamRepository(self)
+        self.document_workflows = ClinicianDocumentWorkflowService(
+            db=db,
+            care_team_repo=self.care_team_repo,
+            execute=lambda *args, **kwargs: self._execute(*args, **kwargs),
+        )
 
     # ── Profile ─────────────────────────────────────
 
@@ -490,19 +495,7 @@ class ClinicianService:
         )
         allergies = cast(list[dict[str, Any]], allergies_res.data or [])
 
-        docs = await self._execute(
-            self.db.table("documents")
-            .select(
-                "id, file_name, document_type, parse_status, ai_summary, "
-                "created_at, uploaded_by_role, clinician_annotation, "
-                "review_status, reviewed_by, reviewed_at, review_note"
-            )
-            .eq("patient_id", pid)
-            .order("created_at", desc=True)
-        )
-        documents = await self._attach_document_reviewers(
-            cast(list[dict[str, Any]], docs.data or [])
-        )
+        documents = await self.document_workflows.fetch_patient_documents(patient_id)
 
         soap_res = await self._execute(
             self.db.table("soap_notes")
@@ -563,48 +556,7 @@ class ClinicianService:
 
     async def list_document_review_queue(self, clinician_id: UUID) -> list[dict[str, Any]]:
         """List pending patient-uploaded documents for assigned patients."""
-        patient_ids = await self._get_assigned_patient_ids(clinician_id)
-        if not patient_ids:
-            return []
-
-        patient_id_values = [str(patient_id) for patient_id in patient_ids]
-        patient_rows = await self._execute(
-            self.db.table("patients")
-            .select("id, first_name, last_name")
-            .in_("id", patient_id_values)
-        )
-        patients_by_id = {
-            str(row["id"]): row
-            for row in cast(list[dict[str, Any]], patient_rows.data or [])
-            if row.get("id")
-        }
-
-        docs = await self._execute(
-            self.db.table("documents")
-            .select(
-                "id, patient_id, file_name, document_type, parse_status, ai_summary, "
-                "source_clinic, created_at, uploaded_by_role, review_status"
-            )
-            .in_("patient_id", patient_id_values)
-            .eq("uploaded_by_role", UploaderRole.PATIENT.value)
-            .eq("review_status", DocumentReviewStatus.PENDING.value)
-            .order("created_at", desc=True)
-        )
-
-        queue_items: list[dict[str, Any]] = []
-        for row in cast(list[dict[str, Any]], docs.data or []):
-            patient = patients_by_id.get(str(row.get("patient_id")))
-            if not patient:
-                continue
-            queue_items.append(
-                {
-                    **row,
-                    "patient_first_name": patient.get("first_name", ""),
-                    "patient_last_name": patient.get("last_name", ""),
-                }
-            )
-
-        return queue_items
+        return await self.document_workflows.list_document_review_queue(clinician_id)
 
     async def approve_document_review(
         self,
@@ -613,12 +565,10 @@ class ClinicianService:
         document_id: UUID,
     ) -> dict[str, Any]:
         """Approve a pending patient-uploaded document review."""
-        return await self._set_document_review(
-            clinician_id=clinician_id,
-            patient_id=patient_id,
-            document_id=document_id,
-            review_status=DocumentReviewStatus.APPROVED,
-            review_note=None,
+        return await self.document_workflows.approve_document_review(
+            clinician_id,
+            patient_id,
+            document_id,
         )
 
     async def reject_document_review(
@@ -629,12 +579,11 @@ class ClinicianService:
         review_note: str | None = None,
     ) -> dict[str, Any]:
         """Reject a pending patient-uploaded document review with optional note."""
-        return await self._set_document_review(
-            clinician_id=clinician_id,
-            patient_id=patient_id,
-            document_id=document_id,
-            review_status=DocumentReviewStatus.REJECTED,
-            review_note=review_note,
+        return await self.document_workflows.reject_document_review(
+            clinician_id,
+            patient_id,
+            document_id,
+            review_note,
         )
 
     async def set_patient_obligation(
@@ -673,22 +622,12 @@ class ClinicianService:
 
         Security: verifies both care_team assignment AND document belongs to patient.
         """
-        await self._assert_patient_assignment(clinician_id, patient_id)
-
-        await self._get_document_row(patient_id, document_id)
-
-        await self._execute(
-            self.db.table("documents")
-            .update(
-                {
-                    "clinician_annotation": annotation_text,
-                    "annotation_by": str(clinician_id),
-                }
-            )
-            .eq("id", str(document_id))
+        return await self.document_workflows.save_document_annotation(
+            clinician_id,
+            patient_id,
+            document_id,
+            annotation_text,
         )
-
-        return {"status": "saved", "document_id": str(document_id)}
 
     async def get_patient_a2a_timeline(
         self,
@@ -739,109 +678,6 @@ class ClinicianService:
         )
         if not assignment_rows:
             raise AuthorizationError("You are not assigned to this patient")
-
-    async def _get_document_row(
-        self,
-        patient_id: UUID,
-        document_id: UUID,
-        *,
-        fields: str = (
-            "id, patient_id, uploaded_by_role, review_status, reviewed_by, "
-            "reviewed_at, review_note"
-        ),
-    ) -> dict[str, Any]:
-        """Fetch a patient-owned document row or raise NotFound."""
-        doc_result = await self._execute(
-            self.db.table("documents")
-            .select(fields)
-            .eq("id", str(document_id))
-            .eq("patient_id", str(patient_id))
-            .single()
-        )
-        if not doc_result.data:
-            raise NotFoundError("Document", str(document_id))
-        return cast(dict[str, Any], doc_result.data)
-
-    async def _set_document_review(
-        self,
-        *,
-        clinician_id: UUID,
-        patient_id: UUID,
-        document_id: UUID,
-        review_status: DocumentReviewStatus,
-        review_note: str | None,
-    ) -> dict[str, Any]:
-        """Persist a shared review decision for a patient-uploaded document."""
-        await self._assert_patient_assignment(clinician_id, patient_id)
-        document = await self._get_document_row(patient_id, document_id)
-
-        if document.get("uploaded_by_role") != UploaderRole.PATIENT.value:
-            raise ValidationError("Only patient-uploaded documents can be reviewed")
-
-        current_status = document.get("review_status")
-        if current_status != DocumentReviewStatus.PENDING.value:
-            raise ValidationError("Document review has already been completed")
-
-        reviewed_at = datetime.now(UTC).isoformat()
-        update_payload = {
-            "review_status": review_status.value,
-            "reviewed_by": str(clinician_id),
-            "reviewed_at": reviewed_at,
-            "review_note": review_note.strip() if review_note else None,
-        }
-
-        result = await self._execute(
-            self.db.table("documents")
-            .update(update_payload)
-            .eq("id", str(document_id))
-            .eq("patient_id", str(patient_id))
-        )
-        updated_rows = cast(list[dict[str, Any]], result.data or [])
-        updated = updated_rows[0] if updated_rows else {**document, **update_payload}
-
-        return {
-            "status": "reviewed",
-            "document_id": str(document_id),
-            "patient_id": str(patient_id),
-            "review_status": updated.get("review_status", review_status.value),
-            "reviewed_by": updated.get("reviewed_by", str(clinician_id)),
-            "reviewed_at": updated.get("reviewed_at", reviewed_at),
-            "review_note": updated.get("review_note"),
-        }
-
-    async def _attach_document_reviewers(
-        self,
-        documents: list[dict[str, Any]],
-    ) -> list[dict[str, Any]]:
-        """Hydrate reviewer metadata onto document rows for clinician views."""
-        reviewer_ids = sorted(
-            {
-                str(document["reviewed_by"])
-                for document in documents
-                if document.get("reviewed_by")
-            }
-        )
-        if not reviewer_ids:
-            return documents
-
-        reviewers_result = await self._execute(
-            self.db.table("clinicians")
-            .select("id, first_name, last_name")
-            .in_("id", reviewer_ids)
-        )
-        reviewers_by_id = {
-            str(row["id"]): row
-            for row in cast(list[dict[str, Any]], reviewers_result.data or [])
-            if row.get("id")
-        }
-
-        for document in documents:
-            reviewer_id = document.get("reviewed_by")
-            document["reviewer"] = (
-                reviewers_by_id.get(str(reviewer_id)) if reviewer_id else None
-            )
-
-        return documents
 
     async def _get_pending_medwatch_count(self, patient_ids: list[UUID]) -> int:
         """Count draft/open MedWatch assessments across all assigned patients."""
